UPSTREAM: tpm: Remove USER_TPMx options, make TPM1/TPM2 menuconfig visible
We would like to have an easy way to completely disable TPM support on a
board. For boards that don't pre-select a TPM protocol via the
MAINBOARD_HAS_TPMx options, this is already possible with the
USER_NO_TPM option. In order to make this available for all boards, this
patch just removes the whole USER_TPMx option group and directly makes
the TPM1 and TPM2 options visible to menuconfig. The MAINBOARD_HAS_TPMx
options can still be used to select defaults and to prevent selection of
a protocol that the TPM is known to not support, but the NO_TPM option
always remains available.

Also fix some mainboards that selected TPM2 directly, which they're not
supposed to do (that's what MAINBOARD_HAS_TPM2 is for), and add a
missing dependency to TPM_CR50 so it is set correctly for a NO_TPM
scenario.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ic461c10ac4c9e168415345ac3d7905b7e62442f2
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 8ad93797d6e1eb2d4be4010e29152636551567fa
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Change-Id: Ib0a73da3c42fa4e8deffecb53f29ee38cbb51a93
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/54641
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Original-Reviewed-by: Christian Walter <christian.walter@9elements.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2927536
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>