UPSTREAM: drivers/i2c/rx6110sa: Add a Kconfig switch to disable ACPI support
In commit b64db833d6 a basic ACPI support was added to the driver.
With this support an SSDT-entry is created for this RTC and it is now
visible to the OS via ACPI. In Linux the PNP-devices, which are
reported over ACPI, are scanned rather early and if the entry is found,
the device is claimed even if there is no driver available yet.
In this case, when the native RTC-driver without ACPI-support is loaded
and tries to register this device, the RTC is already blocked by the
PNP-drivers and cannot be used anymore. This leads to a non-usable RTC
on kernels where the needed ACPI-extension is not yet merged into the
RTC driver.

This patch provides a way to disable the ACPI-support for the RTC if
needed.

BUG=none
BRANCH=none
TEST=none

Change-Id: I33a568ff0faa9c741ee415f23d8e6a9e14da5fef
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 964948d97f1c47769d29622e0a238c74cba32edc
Original-Change-Id: Ic65794d409d13a78d17275c86ec14ee6f04cd2a6
Original-Signed-off-by: Werner Zeh <werner.zeh@siemens.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/55003
Original-Reviewed-by: Mario Scheithauer <mario.scheithauer@siemens.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2927570
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>