UPSTREAM: mb/siemens/mc_apl{1,2,3,5,6}: Disable ACPI-support for RX6110
Already released Linux versions did not have the needed ACPI-extension
in the RTC driver. If the ACPI-Support is enabled for the RTC, this
older Linux will not be able to use this device as it will be claimed by
the PNP-drivers. As there is no way to avoid that an older Linux kernel
meets a newer coreboot in the field, we need to disable the ACPI
support for the RTC for the mc_apl-based mainboards.

BUG=none
BRANCH=none
TEST=none

Change-Id: I85585098950ed49f88ee714a7d671062f22ac923
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 80d5a05cfd08936d5a5f2c44643358d388d84d63
Original-Change-Id: I9f9939ba3234dc3654a4ef8a498649453941ebdf
Original-Signed-off-by: Werner Zeh <werner.zeh@siemens.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/55004
Original-Reviewed-by: Mario Scheithauer <mario.scheithauer@siemens.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2927571
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>