soc/intel/{apl,cnl}: Remove FSP CAR option
One of the reason FSP-T support had to be kept in place was for
Intel Bootguard. This now works with native CAR code, so there is no
reason to keep FSP-T as an option for these platforms.

APL did not even build with FSP_CAR and finding FSP-T using walkcbfs
was only recently fixed using FMAP, so there can be no doubt that this
option was never used with coreboot master.

Change-Id: I0d5844b5a6fd291a13e5f467f4fc682b17eafa63
Signed-off-by: Arthur Heymans <arthur@aheymans.xyz>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/55518
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Ignore-CL-Reviewed-on: https://review.coreboot.org/c/coreboot/+/55026
Ignore-CL-Reviewed-on: https://review.coreboot.org/c/coreboot/+/55228
Ignore-CL-Reviewed-on: https://review.coreboot.org/c/coreboot/+/55304
Ignore-CL-Reviewed-on: https://review.coreboot.org/c/coreboot/+/55280
Ignore-CL-Reviewed-on: https://review.coreboot.org/c/coreboot/+/55414
Ignore-CL-Reviewed-on: https://review.coreboot.org/c/coreboot/+/55497
Ignore-CL-Reviewed-on: https://review.coreboot.org/c/coreboot/+/55498
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2976375
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>