UPSTREAM: soc/intel/common: Fix bugs for GPIO_LOCK_UNLOCK
Per the Intel External Design Specification (doc #618876), the opcode
for GPIO_LOCK_UNLOCK is 0x13.  This commit fixes a bug where the opcode
was defined as 13 decimal instead of hexadecimal.  Additionally, it
fixes another issue where the `pcr_execute_sideband_msg()` function
doesn't actually write the data when this opcode is selected.

BUG=b:191189275
BRANCH=None
TEST=With additional code that uses this opcode, verify that the lock
functionality works by locking a pad in firmware and attempting to
modify the configuration of the pad from the OS.

Signed-off-by: Karthikeyan Ramasubramanian <kramasub@google.com>
Original-Commit-Id: 193ee64d52a707809982231cd857329fe17f9f99
Original-Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Original-Change-Id: Ie14fff595474cdfd647c2b36f1eeb5e018f67375
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/55556
Original-Reviewed-by: Karthik Ramasubramanian <kramasub@google.com>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Change-Id: Ie870a8876b68f6e4e5a984e4a99649d3758b80af
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2973799
Reviewed-by: Karthikeyan Ramasubramanian <kramasub@chromium.org>
Tested-by: Karthikeyan Ramasubramanian <kramasub@chromium.org>
Commit-Queue: Aseda Aboagye <aaboagye@chromium.org>