UPSTREAM: soc/intel/common/block/gpio: Add `gpio_lock_pad()`
This commit adds a method for locking a GPIO pad configuration and its
TX state.  When the configuration is locked, the following registers
become Read-Only and software writes to these registers have no effect.

  Pad Configuration registers
  GPI_NMI_EN
  GPI_SMI_EN
  GPI_GPE_EN

Note that this is only effective if the pad is owned by the host (set in
the PAD_OWN register).

Intel platforms that wish to leverage this function need to define the
PADCFGLOCK offset for their platform.

BUG=b:191189275
BRANCH=None
TEST=With some other code, call gpio_lock_pad() against a pad and verify
that the pad configuration is locked and the state of the pad cannot be
changed from the OS.

Signed-off-by: Karthikeyan Ramasubramanian <kramasub@google.com>
Original-Commit-Id: e58e6f2adfb65fb960cdc41289a5186b4370fd1e
Original-Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Original-Change-Id: Id3c0da2f6942099c0289ca1e33a33c176f49d380
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/55557
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Original-Reviewed-by: Karthik Ramasubramanian <kramasub@google.com>
Change-Id: I94ea9b27c77cd946fa7def055ce725a19afe8ef8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2973800
Reviewed-by: Karthikeyan Ramasubramanian <kramasub@chromium.org>
Tested-by: Karthikeyan Ramasubramanian <kramasub@chromium.org>
Commit-Queue: Aseda Aboagye <aaboagye@chromium.org>